refactor step 5 - modules into rendering
