+++ app/src/main/java/com/replica/replicaisland/AnimationComponent.kt
@@ -23,6 +23,7 @@ import com.replica.replicaisland.core.BaseObject
 import com.replica.replicaisland.core.GameObject
 import com.replica.replicaisland.entities.PlayerComponent
 import com.replica.replicaisland.mechanics.HotSpotSystem
+import com.replica.replicaisland.rendering.SpriteComponent
 import com.replica.replicaisland.sound.SoundSystem
 import com.replica.replicaisland.sound.SoundSystem.Sound
 import kotlin.math.abs
