+++ app/src/main/java/com/replica/replicaisland/AnimationFrame.kt
@@ -16,6 +16,7 @@
 package com.replica.replicaisland
 
 import com.replica.replicaisland.mechanics.CollisionVolume
+import com.replica.replicaisland.rendering.Texture
 
 /**
  * A single animation frame.  Frames contain a texture, a hold time, and collision volumes to
