+++ app/src/main/java/com/replica/replicaisland/FadeDrawableComponent.kt
@@ -21,7 +21,9 @@ import com.replica.replicaisland.Lerp.ease
 import com.replica.replicaisland.Lerp.lerp
 import com.replica.replicaisland.core.BaseObject
 import com.replica.replicaisland.core.GameObject
+import com.replica.replicaisland.rendering.DrawableBitmap
 import com.replica.replicaisland.rendering.RenderComponent
+import com.replica.replicaisland.rendering.Texture
 
 class FadeDrawableComponent : GameComponent() {
     private var mTexture: Texture? = null
