+++ app/src/main/java/com/replica/replicaisland/FixedAnimationComponent.kt
@@ -19,6 +19,7 @@ package com.replica.replicaisland
 
 import com.replica.replicaisland.core.BaseObject
 import com.replica.replicaisland.core.GameObject
+import com.replica.replicaisland.rendering.SpriteComponent
 
 class FixedAnimationComponent : GameComponent() {
     private var animationIndex = 0
