+++ app/src/main/java/com/replica/replicaisland/FrameRateWatcherComponent.kt
@@ -16,6 +16,7 @@
 package com.replica.replicaisland
 
 import com.replica.replicaisland.core.BaseObject
+import com.replica.replicaisland.rendering.DrawableObject
 import com.replica.replicaisland.rendering.RenderComponent
 
 class FrameRateWatcherComponent : GameComponent() {
