+++ app/src/main/java/com/replica/replicaisland/GLSurfaceView.kt
@@ -24,6 +24,7 @@ import android.util.AttributeSet
 import android.view.SurfaceHolder
 import android.view.SurfaceView
 import com.replica.replicaisland.GLSurfaceView.Renderer
+import com.replica.replicaisland.rendering.TextureLibrary
 import java.io.Writer
 import java.util.*
 import javax.microedition.khronos.egl.*
