+++ app/src/main/java/com/replica/replicaisland/GameThread.kt
@@ -19,6 +19,7 @@ package com.replica.replicaisland
 
 import android.os.SystemClock
 import com.replica.replicaisland.core.BaseObject
+import com.replica.replicaisland.rendering.GameRenderer
 
 /**
  * The GameThread contains the main loop for the game engine logic.  It invokes the game graph,
