+++ app/src/main/java/com/replica/replicaisland/GenericAnimationComponent.kt
@@ -18,6 +18,7 @@ package com.replica.replicaisland
 import com.replica.replicaisland.core.GameObject.ActionType
 import com.replica.replicaisland.core.BaseObject
 import com.replica.replicaisland.core.GameObject
+import com.replica.replicaisland.rendering.SpriteComponent
 
 class GenericAnimationComponent : GameComponent() {
     private var mSprite: SpriteComponent? = null
