+++ app/src/main/java/com/replica/replicaisland/GhostComponent.kt
@@ -21,6 +21,7 @@ import com.replica.replicaisland.core.GameObject.ActionType
 import com.replica.replicaisland.core.BaseObject
 import com.replica.replicaisland.core.GameObject
 import com.replica.replicaisland.entities.PlayerComponent
+import com.replica.replicaisland.rendering.SpriteComponent
 import com.replica.replicaisland.sound.SoundSystem
 import com.replica.replicaisland.sound.SoundSystem.Sound
