+++ app/src/main/java/com/replica/replicaisland/ObjectRegistry.kt
@@ -27,7 +27,10 @@ import com.replica.replicaisland.mechanics.EventRecorder
 import com.replica.replicaisland.mechanics.HotSpotSystem
 import com.replica.replicaisland.mechanics.TimeSystem
 import com.replica.replicaisland.rendering.CameraSystem
+import com.replica.replicaisland.rendering.DrawableFactory
+import com.replica.replicaisland.rendering.OpenGLSystem
 import com.replica.replicaisland.rendering.RenderSystem
+import com.replica.replicaisland.rendering.TextureLibrary
 import com.replica.replicaisland.sound.SoundSystem
 import java.util.*
