+++ app/src/main/java/com/replica/replicaisland/entities/DoorAnimationComponent.kt
@@ -3,7 +3,7 @@ package com.replica.replicaisland.entities
 import com.replica.replicaisland.mechanics.ChannelSystem
 import com.replica.replicaisland.GameComponent
 import com.replica.replicaisland.SolidSurfaceComponent
-import com.replica.replicaisland.SpriteComponent
+import com.replica.replicaisland.rendering.SpriteComponent
 import com.replica.replicaisland.core.BaseObject
 import com.replica.replicaisland.core.GameObject
 import com.replica.replicaisland.sound.SoundSystem
