+++ app/src/main/java/com/replica/replicaisland/entities/NPCAnimationComponent.kt
@@ -2,7 +2,7 @@ package com.replica.replicaisland.entities
 
 import com.replica.replicaisland.mechanics.ChannelSystem
 import com.replica.replicaisland.GameComponent
-import com.replica.replicaisland.SpriteComponent
+import com.replica.replicaisland.rendering.SpriteComponent
 import com.replica.replicaisland.core.BaseObject
 import com.replica.replicaisland.core.GameObject
 import kotlin.math.abs
