+++ app/src/main/java/com/replica/replicaisland/levels/LevelBuilder.kt
@@ -7,7 +7,7 @@ import com.replica.replicaisland.R
 import com.replica.replicaisland.rendering.RenderComponent
 import com.replica.replicaisland.ScrollerComponent
 import com.replica.replicaisland.SortConstants
-import com.replica.replicaisland.TiledVertexGrid
+import com.replica.replicaisland.rendering.TiledVertexGrid
 import kotlin.math.max
 import kotlin.math.min
