+++ app/src/main/java/com/replica/replicaisland/rendering/DrawableObject.kt
@@ -1,19 +1,8 @@
-/*
- * Copyright (C) 2010 The Android Open Source Project
- * Copyright (C) 2025 Jim Andreas kotlin conversion
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package com.replica.replicaisland
+package com.replica.replicaisland.rendering
+
+import com.replica.replicaisland.AllocationGuard
+import com.replica.replicaisland.ObjectPool
+import com.replica.replicaisland.Vector2
 
 /**
  * DrawableObject is the base object interface for objects that can be rendered to the screen.
