+++ app/src/main/java/com/replica/replicaisland/rendering/OpenGLSystem.kt
@@ -1,21 +1,4 @@
-/*
- * Copyright (C) 2010 The Android Open Source Project
- * Copyright (C) 2025 Jim Andreas kotlin conversion
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-@file:Suppress("unused")
-
-package com.replica.replicaisland
+package com.replica.replicaisland.rendering
 
 import com.replica.replicaisland.core.BaseObject
 import javax.microedition.khronos.opengles.GL10
@@ -64,7 +47,8 @@ class OpenGLSystem : BaseObject {
             var cropSignature = crop[0] + crop[1] shl 16
             cropSignature = cropSignature or crop[2] + crop[3]
             if (cropSignature != sLastSetCropSignature) {
-                (sGL as GL11?)!!.glTexParameteriv(GL10.GL_TEXTURE_2D, GL11Ext.GL_TEXTURE_CROP_RECT_OES,
+                (sGL as GL11?)!!.glTexParameteriv(
+                    GL10.GL_TEXTURE_2D, GL11Ext.GL_TEXTURE_CROP_RECT_OES,
                         crop, 0)
                 sLastSetCropSignature = cropSignature
             }
