+++ app/src/main/java/com/replica/replicaisland/rendering/SpriteComponent.kt
@@ -1,27 +1,11 @@
-/*
- * Copyright (C) 2010 The Android Open Source Project
- * Copyright (C) 2025 Jim Andreas kotlin conversion
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-@file:Suppress("SENSELESS_COMPARISON")
+package com.replica.replicaisland.rendering
 
-package com.replica.replicaisland
-
+import com.replica.replicaisland.GameComponent
+import com.replica.replicaisland.rendering.SpriteAnimation
 import com.replica.replicaisland.core.BaseObject
 import com.replica.replicaisland.core.GameObject
 import com.replica.replicaisland.core.PhasedObjectManager
 import com.replica.replicaisland.mechanics.DynamicCollisionComponent
-import com.replica.replicaisland.rendering.RenderComponent
 
 /**
  * Provides an interface for controlling a sprite with animations.  Manages a list of animations
